rtp: move rewriting logic in its own module
